# src/docs/swagger.yaml
openapi: 3.0.0
info:
    title: Registry API
    version: 1.0.0
    description: |
        REST API for managing organizations, their roles, element fields
        and agreement fields within Registry API.

        Each organization is the main container in the system: it groups elements,
        defines the vocabulary they must follow and establishes the available roles.
servers:
    - url: /
...
paths:
...
    /api/v1/users:
        get:
            tags:
                - Users
            summary: List all users
            operationId: getUsers
            security:
                - BearerAuth: []
            description: |
                Returns an array with all users in the system.
                Requires admin role to access.
            responses:
                '200':
                    description: A list of users
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              type: array
                                              items:
                                                  $ref: '#/components/schemas/User'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '500':
                    $ref: '#/components/responses/InternalError'
        post:
            tags:
                - Users
            summary: Create a new user
            operationId: createUser
            description: |
                Creates a new user with the provided credentials.
                Requires admin role to create users.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserInput'
            responses:
                '201':
                    description: User created successfully
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/User'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '409':
                    $ref: '#/components/responses/Conflict'
                '500':
                    $ref: '#/components/responses/InternalError'
    /api/v1/users/{username}:
        parameters:
            - in: path
              name: username
              required: true
              schema:
                  type: string
              description: Username of the user
              example: johndoe
        get:
            tags:
                - Users
            summary: Get user by username
            operationId: getUserByUsername
            security:
                - BearerAuth: []
            description: |
                Returns a specific user by their username.
                Requires admin role to access.
            responses:
                '200':
                    description: User found
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/User'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
        put:
            tags:
                - Users
            summary: Update user by username
            operationId: updateUser
            security:
                - BearerAuth: []
            description: |
                Updates an existing user's information.
                Requires admin role to update users.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserInput'
            responses:
                '200':
                    description: User updated successfully
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/User'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
        delete:
            tags:
                - Users
            summary: Delete user by username
            operationId: deleteUser
            security:
                - BearerAuth: []
            description: |
                Deletes a user from the system.
                Requires admin role to delete users.
            responses:
                '200':
                    description: User deleted successfully
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              type: object
                                              nullable: true
                                              example: null
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
...
components:
...
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: JWT token obtained via POST /api/v1/users/login

    # ── Parameters ────────────────────────────────
...
    responses:
        ValidationError:
            description: Validation error in the submitted data
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Validation error
                        httpStatus: 400
                        appCode: VALIDATION_ERROR
                        data: null
                        error:
                            message: Validation error
                            details: null
        NotFound:
            description: The requested resource does not exist
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Resource not found
                        httpStatus: 404
                        appCode: NOT_FOUND
                        data: null
                        error:
                            message: Resource not found
                            details: null
        Conflict:
            description: A resource with that identifier already exists
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Duplicate key error
                        httpStatus: 409
                        appCode: DUPLICATE_KEY
                        data: null
                        error:
                            message: Duplicate key error
                            details: null
...
        Unauthorized:
            description: Missing or invalid authentication token
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Authorization header missing or malformed
                        httpStatus: 401
                        appCode: UNAUTHORIZED
                        data: null
                        error:
                            message: Authorization header missing or malformed
                            details: null
        Forbidden:
            description: Insufficient permissions to perform this action
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: You do not have permission to perform that action
                        httpStatus: 403
                        appCode: FORBIDDEN
                        data: null
                        error:
                            message: You do not have permission to perform that action
                            details: null
        InternalError:
            description: Internal server error
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Internal server error
                        httpStatus: 500
                        appCode: UNKNOWN_ERROR
                        data: null
                        error:
                            message: Internal server error
                            details: null

    # ── Schemas ───────────────────────────────────
    schemas:
...
        SuccessResponse:
            type: object
            description: Standard wrapper for all successful API responses
            properties:
                success:
                    type: boolean
                    example: true
                message:
                    type: string
                    example: OK
                httpStatus:
                    type: integer
                    example: 200
                appCode:
                    type: string
                    example: SUCCESS
                error:
                    type: object
                    nullable: true
                    example: null
            required:
                - success
                - message
                - httpStatus
                - appCode
                - data
                - error
        ErrorResponse:
            type: object
            description: Standard wrapper for all error API responses
            properties:
                success:
                    type: boolean
                    example: false
                message:
                    type: string
                httpStatus:
                    type: integer
                appCode:
                    type: string
                data:
                    type: object
                    nullable: true
                    example: null
                error:
                    type: object
                    properties:
                        message:
                            type: string
                        details:
                            description: Additional error information (may be null)
                            type: object
                            nullable: true
            required:
                - success
                - message
                - httpStatus
                - appCode
                - data
                - error

        # ── Organization ─────────────────────────
...
        User:
            type: object
            description: Represents a user in the system
            required:
                - _id
                - username
                - email
                - name
                - surname
                - systemRole
            properties:
                _id:
                    type: string
                    description: MongoDB ObjectId
                    example: 665d4e5f6a7b8c9d0e1f2a3b
                username:
                    type: string
                    description: Unique username (3-50 characters)
                    minLength: 3
                    maxLength: 50
                    example: johndoe
                email:
                    type: string
                    format: email
                    description: User email address
                    example: john@example.com
                name:
                    type: string
                    description: User's first name (2-30 characters)
                    minLength: 2
                    maxLength: 30
                    example: John
                surname:
                    type: string
                    description: User's last name (2-50 characters)
                    minLength: 2
                    maxLength: 50
                    example: Doe
                systemRole:
                    type: string
                    description: System-level role
                    enum: [ADMIN, USER]
                    example: ADMIN
                createdAt:
                    type: string
                    format: date-time
                    example: '2025-06-01T12:00:00.000Z'
                updatedAt:
                    type: string
                    format: date-time
                    example: '2025-06-01T12:00:00.000Z'
        UserInput:
            type: object
            description: Request payload for creating or updating a user
            required:
                - username
                - email
                - password
                - name
                - surname
                - systemRole
            properties:
                username:
                    type: string
                    description: Unique username (3-50 characters)
                    minLength: 3
                    maxLength: 50
                    example: johndoe
                email:
                    type: string
                    format: email
                    description: Valid email address
                    example: john@example.com
                password:
                    type: string
                    description: Password (minimum 6 characters)
                    minLength: 6
                    example: securePassword123
                name:
                    type: string
                    description: User's first name (2-30 characters)
                    minLength: 2
                    maxLength: 30
                    example: John
                surname:
                    type: string
                    description: User's last name (2-50 characters)
                    minLength: 2
                    maxLength: 50
                    example: Doe
                systemRole:
                    type: string
                    description: System-level role
                    enum: [ADMIN, USER]
                    example: USER
...